+++ src/main/java/org/sspoid/sspoid/common/resolver/CurrentUserArgumentResolver.java
@@ -8,7 +8,6 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import org.sspoid.sspoid.api.controller.CurrentUser;
 import org.sspoid.sspoid.common.security.JwtTokenProvider;
 
 @Component
